refactor: extract tool execution into tool-execution.js (#147)
Brings tool-loop-handler.js from 1169 to 774 lines (under the 1000-line
cap). Moves _executeToolCalls, _parseToolCallArguments,
_recordInvalidArgsResult, _promptToolConfirmation, _truncateInitialResult
and the tool-justification store verbatim; preserves all existing lint
suppression state and the 'ToolLoop' logger name for log continuity.
chat-handler.js now imports retrieveToolJustification from the new
module. Drops the tool-loop-handler entry from the file-size baseline.
