+++ tests/baselines/file-sizes.json
@@ -5,10 +5,6 @@
       "lines": 1636
     },
     {
-      "path": "scripts/core/tool-loop-handler.js",
-      "lines": 1057
-    },
-    {
       "path": "scripts/ui/simulacrum-sidebar-tab.js",
       "lines": 1489
     },
